requirements.txt
dk
Django==1.7.11
-e .
py
pytest
pytest-cov
pytest-xdist
django-localflavor==1.3
